data/bakajstep/KNN_Project2024/jobspec-cfg/requirements.txt
datasets==2.18.0
accelerate==0.30.0
conllu==4.5.3
evaluate==0.4.2
numpy==1.24.3
nltk~=3.8.1
Requests==2.31.0
pandas==2.2.2
scikit_learn==1.1.3
torch==2.0.0
transformers==4.37.2
pagexml-tools==0.5.0
evaluate~=0.4.2
seqeval~=1.2.2
PyYAML~=6.0.1
tqdm==4.65.0
nltk~=3.8.1
